# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": " Questioner helps meetup organizer prioritize questions and answers. ",
    "version": "1.0.0",
    "title": "Questioner",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "info@questioner.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "questioner1.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/meetups/{meetupId}": {
      "get": {
        "tags": [
          "Meetups"
        ],
        "summary": "Gets Meetup by Id",
        "description": "",
        "operationId": "getMeetupById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "meetupId",
            "in": "path",
            "description": "Gets meetup by id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication Errors"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Meetups"
        ],
        "summary": "Deletes a Meetup",
        "description": "",
        "operationId": "deleteMeetup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "meetupId",
            "in": "path",
            "description": "meetup id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation Or Meetup-Id Doesnt Exit"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication Errors"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": []
          }
        ]
      }
    },
...
    "/questions/{questionId}/downvote": {
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Down vote on a question",
        "description": "",
        "operationId": "downVote",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation Or Meetup-Id Doesnt Exit"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication Errors"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "tokens": {
      "type": "apiKey",
      "name": "token",
      "in": "header"
    }
  },
...
}